+++ app/src/main/java/com/github/sdpsharelook/MainActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import android.view.View
 import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
-import com.github.sdpsharelook.Section.SectionActivity
+import com.github.sdpsharelook.section.SectionActivity
 
 import com.github.sdpsharelook.textDetection.TextDetectionActivity
