+++ app/src/main/java/com/github/sdpsharelook/section/CardAdapter.kt
@@ -1,13 +1,10 @@
-package com.github.sdpsharelook.Section
+package com.github.sdpsharelook.section
 
 import android.app.Dialog
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.github.sdpsharelook.R
-import com.github.sdpsharelook.databinding.ActivitySectionBinding
 import com.github.sdpsharelook.databinding.CardSectionBinding
-import com.github.sdpsharelook.databinding.PopupBinding
 
 class CardAdapter(
     private val sections: List<Section>,
