+++ app/src/main/java/com/github/sdpsharelook/section/CardViewHolder.kt
@@ -1,8 +1,7 @@
-package com.github.sdpsharelook.Section
+package com.github.sdpsharelook.section
 
 import androidx.recyclerview.widget.RecyclerView
 import com.github.sdpsharelook.databinding.CardSectionBinding
-import java.lang.ref.WeakReference
 
 
 class CardViewHolder(
