+++ app/src/main/java/com/github/sdpsharelook/section/CountryItem.kt
@@ -1,4 +1,4 @@
-package com.github.sdpsharelook.Section
+package com.github.sdpsharelook.section
 
 
 class CountryItem (
