+++ app/src/main/java/com/github/sdpsharelook/section/Section.kt
@@ -1,4 +1,4 @@
-package com.github.sdpsharelook.Section
+package com.github.sdpsharelook.section
 
 var sectionList = mutableListOf<Section>()
