+++ app/src/main/java/com/github/sdpsharelook/section/SectionActivity.kt
@@ -1,4 +1,4 @@
-package com.github.sdpsharelook.Section
+package com.github.sdpsharelook.section
 
 import android.app.Dialog
 import android.content.Intent
