+++ app/src/main/java/com/github/sdpsharelook/section/SectionClickListener.kt
@@ -1,4 +1,4 @@
-package com.github.sdpsharelook.Section
+package com.github.sdpsharelook.section
 
 interface SectionClickListener {
     fun onClick(section: Section)
