+++ app/src/main/java/com/github/sdpsharelook/section/SectionDetail.kt
@@ -1,4 +1,4 @@
-package com.github.sdpsharelook.Section
+package com.github.sdpsharelook.section
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
