+++ app/src/test/java/com/github/sdpsharelook/section/SectionActivityTest.kt
@@ -1,4 +1,4 @@
-package com.github.sdpsharelook
+package com.github.sdpsharelook.section
 
 
 import android.view.View
@@ -10,7 +10,7 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import com.github.sdpsharelook.Section.SectionActivity
+import com.github.sdpsharelook.R
 import org.hamcrest.Description
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
